requirements.txt
pyspark==3.5.1
delta-spark==3.2.0
scikit-learn==1.4.2
matplotlib==3.8.4
pandas==2.2.2
numpy==1.26.4
